+++ src/webview/readline.ts
@@ -1,4 +1,4 @@
-import { LineInputModel, ModelEdit, EditableDocument, emptySelectionOption } from "../cursor-doc/model";
+import { LineInputModel, ModelEdit, EditableDocument, ModelEditSelection } from "../cursor-doc/model";
 import { Token, validPair } from "../cursor-doc/clojure-lexer";
 import { TokenCursor, LispTokenCursor } from "../cursor-doc/token-cursor";
 
@@ -44,38 +44,39 @@ export class ReplReadline implements EditableDocument {
     }
 
     /** The offset of the start of the selection into the document. */
-    private _selectionStart: number = 0;
+    //private _selectionStart: number = 0;
+
+    private _selection = new ModelEditSelection(0);
 
     /** Returns the offset of the start of the selection. */
-    get selectionStart() {
-        return this._selectionStart
+    get selectionLeft() {
+        return this._selection.anchor;
     };
 
-    /** Sets the start of the selection. */
-    set selectionStart(val: number) {
-        this._selectionStart = Math.min(this.model.maxOffset, Math.max(val, 0));
+    // /** Sets the start of the selection. */
+    set selectionLeft(val: number) {
+        this._selection.anchor = Math.min(this.model.maxOffset, Math.max(val, 0));
     }
 
-    set selection(selection: { anchor: number, active: number}) {
-        this.selectionStart = selection.anchor;
-        this.selectionEnd = selection.active;
+    set selection(selection: ModelEditSelection) {
+        this._selection = selection;
     }
 
-    get selection(): { anchor: number, active: number} {
-        return { anchor: this.selectionStart, active: this.selectionEnd}
+    get selection(): ModelEditSelection {
+        return new ModelEditSelection(this.selectionLeft, this.selectionRight);
     }
 
     /** The offset of the end of the selection into the document. */
-    private _selectionEnd: number = 0;
+    //private _selectionEnd: number = 0;
 
     /** Returns the offset of the end of the selection. */
-    get selectionEnd() {
-        return this._selectionEnd
+    get selectionRight() {
+        return this._selection.active;
     };
 
     /** Sets the end of the selection. */
-    set selectionEnd(val: number) {
-        this._selectionEnd = Math.min(this.model.maxOffset, Math.max(val, 0));
+    set selectionRight(val: number) {
+        this._selection.active = Math.min(this.model.maxOffset, Math.max(val, 0));
     }
 
     /** The underlying tokenized source. */
@@ -98,12 +99,12 @@ export class ReplReadline implements EditableDocument {
 
     /**
      * Returns a TokenCursor into the document.
-     * 
+     *
      * @param row the line to position the cursor at.
-     * @param col the column to position the cursor at. 
+     * @param col the column to position the cursor at.
      * @param previous if true, position the cursor at the previous token.
      */
-    public getTokenCursor(offset: number = this.selectionEnd, previous: boolean = false) {
+    public getTokenCursor(offset: number = this.selectionRight, previous: boolean = false) {
         let [row, col] = this.model.getRowCol(offset);
         let line = this.model.lines[row]
         let lastIndex = 0;
@@ -121,9 +122,9 @@ export class ReplReadline implements EditableDocument {
     /**
      * Executes a block of code, during which any edits that are performed on the document will be created with Undo support.
      * This should happen almost all of the time- in fact the only time it shouldn't is when replaying undo/redo operations.
-     * 
+     *
      * FIXME: Perhaps this should be "withoutUndo"?
-     * 
+     *
      * @param body the code to execute.
      */
     withUndo(body: () => void) {
@@ -138,18 +139,18 @@ export class ReplReadline implements EditableDocument {
 
     /**
      * Inserts a string at the current cursor location.
-     * 
+     *
      * @param text the text to insert
      */
     insertString(text: string) {
         this.withUndo(() => {
-            let cs = Math.min(this.selectionStart, this.selectionEnd);
-            let ce = Math.max(this.selectionStart, this.selectionEnd);
+            let cs = Math.min(this.selectionLeft, this.selectionRight);
+            let ce = Math.max(this.selectionLeft, this.selectionRight);
             this.model.edit([
                 new ModelEdit('changeRange', [cs, ce, text, [cs, ce], [cs + text.length, cs + text.length]])
-            ], { selection: emptySelectionOption(cs + text.length) });
+            ], { selection: new ModelEditSelection(cs + text.length) });
             this.repaint();
-            this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+            this.caretX = this.model.getRowCol(this.selectionRight)[1];
         });
     }
 
@@ -159,8 +160,8 @@ export class ReplReadline implements EditableDocument {
     }
 
     maybeShowCompletion() {
-        if (this.getTokenCursor().offsetStart == this.selectionEnd && !this.getTokenCursor().previous().isWhiteSpace()) {
-            let evt: CompletionEvent = { type: "show", position: this.selectionEnd, toplevel: this.model.getText(0, this.model.maxOffset) }
+        if (this.getTokenCursor().offsetStart == this.selectionRight && !this.getTokenCursor().previous().isWhiteSpace()) {
+            let evt: CompletionEvent = { type: "show", position: this.selectionRight, toplevel: this.model.getText(0, this.model.maxOffset) }
             this._completionListeners.forEach(x => x(evt));
         } else
             this.clearCompletion();
@@ -168,202 +169,202 @@ export class ReplReadline implements EditableDocument {
 
     /**
      * Moves the caret left one character, using text editor semantics.
-     * 
+     *
      * @param clear if true, clears the current selection, if any, otherwise moves `cursorEnd` only.
      */
     caretLeft(clear: boolean = true) {
         this.clearCompletion();
-        if (clear && this.selectionStart != this.selectionEnd) {
-            if (this.selectionStart < this.selectionEnd)
-                this.selectionEnd = this.selectionStart;
+        if (clear && this.selectionLeft != this.selectionRight) {
+            if (this.selectionLeft < this.selectionRight)
+                this.selectionRight = this.selectionLeft;
             else
-                this.selectionStart = this.selectionEnd;
+                this.selectionLeft = this.selectionRight;
         } else {
-            this.selectionEnd--;
+            this.selectionRight--;
             if (clear)
-                this.selectionStart = this.selectionEnd;
+                this.selectionLeft = this.selectionRight;
         }
         this.repaint();
-        this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+        this.caretX = this.model.getRowCol(this.selectionRight)[1];
     }
 
     /**
      * Moves the caret right one character, using text editor semantics.
-     * 
+     *
      * @param clear if true, clears the current selection, if any, otherwise moves `cursorEnd` only.
      */
     caretRight(clear: boolean = true) {
         this.clearCompletion();
-        if (clear && this.selectionStart != this.selectionEnd) {
-            if (this.selectionStart > this.selectionEnd)
-                this.selectionEnd = this.selectionStart;
+        if (clear && this.selectionLeft != this.selectionRight) {
+            if (this.selectionLeft > this.selectionRight)
+                this.selectionRight = this.selectionLeft;
             else
-                this.selectionStart = this.selectionEnd;
+                this.selectionLeft = this.selectionRight;
         } else {
-            this.selectionEnd++
+            this.selectionRight++
             if (clear)
-                this.selectionStart = this.selectionEnd;
+                this.selectionLeft = this.selectionRight;
         }
         this.repaint();
-        this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+        this.caretX = this.model.getRowCol(this.selectionRight)[1];
     }
 
     /**
      * Moves the caret to the beginning of the document, using text editor semantics.
-     * 
+     *
      * @param clear if true, clears the current selection, if any, otherwise moves `cursorEnd` only.
      */
     caretHomeAll(clear: boolean = true) {
         this.clearCompletion();
-        this.selectionEnd = 0;
+        this.selectionRight = 0;
         if (clear)
-            this.selectionStart = this.selectionEnd;
+            this.selectionLeft = this.selectionRight;
         this.repaint();
-        this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+        this.caretX = this.model.getRowCol(this.selectionRight)[1];
     }
 
     /**
      * Moves the caret to the end of the document, using text editor semantics.
-     * 
+     *
      * @param clear if true, clears the current selection, if any, otherwise moves `cursorEnd` only.
      */
     caretEndAll(clear: boolean = true) {
         this.clearCompletion();
-        this.selectionEnd = this.model.maxOffset;
+        this.selectionRight = this.model.maxOffset;
         if (clear)
-            this.selectionStart = this.selectionEnd;
+            this.selectionLeft = this.selectionRight;
         this.repaint();
-        this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+        this.caretX = this.model.getRowCol(this.selectionRight)[1];
     }
 
     /**
      * Moves the caret to the beginning of the line, using text editor semantics.
-     * 
+     *
      * @param clear if true, clears the current selection, if any, otherwise moves `cursorEnd` only.
      */
     caretHome(clear: boolean = true) {
         this.clearCompletion();
-        let [row, col] = this.model.getRowCol(this.selectionEnd);
-        this.selectionEnd = this.selectionEnd - col;
+        let [row, col] = this.model.getRowCol(this.selectionRight);
+        this.selectionRight = this.selectionRight - col;
         if (clear)
-            this.selectionStart = this.selectionEnd;
+            this.selectionLeft = this.selectionRight;
         this.repaint();
-        this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+        this.caretX = this.model.getRowCol(this.selectionRight)[1];
     }
 
     /**
      * Moves the caret to the end of the line, using text editor semantics.
-     * 
+     *
      * @param clear if true, clears the current selection, if any, otherwise moves `cursorEnd` only.
      */
     caretEnd(clear: boolean = true) {
         this.clearCompletion();
-        let [row, col] = this.model.getRowCol(this.selectionEnd);
-        this.selectionEnd = this.selectionEnd - col + this.model.lines[row].text.length;
+        let [row, col] = this.model.getRowCol(this.selectionRight);
+        this.selectionRight = this.selectionRight - col + this.model.lines[row].text.length;
         if (clear)
-            this.selectionStart = this.selectionEnd;
+            this.selectionLeft = this.selectionRight;
         this.repaint();
-        this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+        this.caretX = this.model.getRowCol(this.selectionRight)[1];
     }
 
     /**
      * Moves the caret to the previous line, using text editor semantics.
-     * 
+     *
      * @param clear if true, clears the current selection, if any, otherwise moves `cursorEnd` only.
      */
     caretUp(clear: boolean = true) {
         this.clearCompletion();
-        let [row, col] = this.model.getRowCol(this.selectionEnd);
+        let [row, col] = this.model.getRowCol(this.selectionRight);
         if (row > 0) {
             let len = this.model.lines[row - 1].text.length;
-            this.selectionEnd = this.model.getOffsetForLine(row - 1) + Math.min(this.caretX, len);
+            this.selectionRight = this.model.getOffsetForLine(row - 1) + Math.min(this.caretX, len);
         } else {
-            this.selectionEnd = 0;
+            this.selectionRight = 0;
         }
         if (clear)
-            this.selectionStart = this.selectionEnd;
+            this.selectionLeft = this.selectionRight;
         this.repaint();
     }
 
     /**
      * Moves the caret to the next line, using text editor semantics.
-     * 
+     *
      * @param clear if true, clears the current selection, if any, otherwise moves `cursorEnd` only.
      */
     caretDown(clear: boolean = true) {
         this.clearCompletion();
-        let [row, col] = this.model.getRowCol(this.selectionEnd);
+        let [row, col] = this.model.getRowCol(this.selectionRight);
         if (row < this.model.lines.length - 1) {
             let len = this.model.lines[row + 1].text.length;
-            this.selectionEnd = this.model.getOffsetForLine(row + 1) + Math.min(this.caretX, len);
+            this.selectionRight = this.model.getOffsetForLine(row + 1) + Math.min(this.caretX, len);
         } else {
-            this.selectionEnd = this.model.maxOffset;
+            this.selectionRight = this.model.maxOffset;
         }
         if (clear)
-            this.selectionStart = this.selectionEnd;
+            this.selectionLeft = this.selectionRight;
         this.repaint();
     }
 
     /**
      * Deletes the current selection.
-     * 
+     *
      * FIXME: this should just be `changeRange`
      */
     private deleteSelection() {
         this.withUndo(() => {
-            if (this.selectionStart != this.selectionEnd) {
-                this.model.deleteRange(Math.min(this.selectionStart, this.selectionEnd), Math.max(this.selectionStart, this.selectionEnd) - Math.min(this.selectionStart, this.selectionEnd));
-                this.selectionStart = this.selectionEnd = Math.min(this.selectionStart, this.selectionEnd);
+            if (this.selectionLeft != this.selectionRight) {
+                this.model.deleteRange(Math.min(this.selectionLeft, this.selectionRight), Math.max(this.selectionLeft, this.selectionRight) - Math.min(this.selectionLeft, this.selectionRight));
+                this.selectionLeft = this.selectionRight = Math.min(this.selectionLeft, this.selectionRight);
             }
         })
     }
 
     /**
      * Retrieve the current selection as text.
-     * 
+     *
      */
     getSelectionText() {
-        if (this.selectionStart != this.selectionEnd) {
-            return this.model.getText(Math.min(this.selectionStart, this.selectionEnd), Math.max(this.selectionStart, this.selectionEnd))
+        if (this.selectionLeft != this.selectionRight) {
+            return this.model.getText(Math.min(this.selectionLeft, this.selectionRight), Math.max(this.selectionLeft, this.selectionRight))
         }
         return "";
     }
 
     /**
      * If there is no selection- deletes the character to the left of the cursor and moves it back one character.
-     * 
+     *
      * If there is a selection, deletes the selection.
      */
     backspace() {
         this.withUndo(() => {
-            if (this.selectionStart != this.selectionEnd) {
+            if (this.selectionLeft != this.selectionRight) {
                 this.deleteSelection();
             } else {
-                if (this.selectionEnd > 0) {
-                    this.model.deleteRange(this.selectionEnd - 1, 1, [this.selectionStart, this.selectionEnd], [this.selectionEnd - 1, this.selectionEnd - 1]);
-                    this.selectionEnd--;
+                if (this.selectionRight > 0) {
+                    this.model.deleteRange(this.selectionRight - 1, 1, [this.selectionLeft, this.selectionRight], [this.selectionRight - 1, this.selectionRight - 1]);
+                    this.selectionRight--;
                 }
-                this.selectionStart = this.selectionEnd;
+                this.selectionLeft = this.selectionRight;
             }
             this.repaint()
-            this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+            this.caretX = this.model.getRowCol(this.selectionRight)[1];
         });
     }
 
     /**
      * If there is no selection- deletes the character to the right of the cursor.
-     * 
+     *
      * If there is a selection, deletes the selection.
      */
     delete() {
         this.withUndo(() => {
-            if (this.selectionStart != this.selectionEnd) {
+            if (this.selectionLeft != this.selectionRight) {
                 this.deleteSelection();
             } else {
-                this.model.deleteRange(this.selectionEnd, 1);
-                this.selectionStart = this.selectionEnd;
+                this.model.deleteRange(this.selectionRight, 1);
+                this.selectionLeft = this.selectionRight;
             }
-            this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+            this.caretX = this.model.getRowCol(this.selectionRight)[1];
             this.repaint()
         });
     }
@@ -469,8 +470,8 @@ export class ReplReadline implements EditableDocument {
     }
 
     /**
-     * Given a TokenCursor, returns the HTMLElement that is rendered for this token. 
-     * @param cursor 
+     * Given a TokenCursor, returns the HTMLElement that is rendered for this token.
+     * @param cursor
      */
     private getElementForToken(cursor: TokenCursor) {
         if (cursor && this.inputLines[cursor.line])
@@ -527,18 +528,18 @@ export class ReplReadline implements EditableDocument {
         this.model.changedLines.clear();
 
         // reposition the caret
-        let [row, col] = this.model.getRowCol(this.selectionEnd);
+        let [row, col] = this.model.getRowCol(this.selectionRight);
         this.inputLines[row].appendChild(this.caret);
         let style = getComputedStyle(this.inputLines[row]);
         ctx.font = style.fontStyle + " " + style.fontSize + " " + style.fontFamily;
 
         this.caret.style.left = measureText(this.model.lines[row].text.substr(0, col)) + "px";
 
-        let startLine = this.model.getRowCol(Math.min(this.lastSelectionStart, this.lastSelectionEnd, this.selectionStart, this.selectionEnd));
-        let endLine = this.model.getRowCol(Math.max(this.lastSelectionStart, this.lastSelectionEnd, this.selectionStart, this.selectionEnd));
+        let startLine = this.model.getRowCol(Math.min(this.lastSelectionStart, this.lastSelectionEnd, this.selectionLeft, this.selectionRight));
+        let endLine = this.model.getRowCol(Math.max(this.lastSelectionStart, this.lastSelectionEnd, this.selectionLeft, this.selectionRight));
 
-        let cs = this.model.getRowCol(Math.min(this.selectionStart, this.selectionEnd));
-        let ce = this.model.getRowCol(Math.max(this.selectionStart, this.selectionEnd));
+        let cs = this.model.getRowCol(Math.min(this.selectionLeft, this.selectionRight));
+        let ce = this.model.getRowCol(Math.max(this.selectionLeft, this.selectionRight));
 
         let lcs = this.model.getRowCol(Math.min(this.lastSelectionStart, this.lastSelectionEnd));
         let lce = this.model.getRowCol(Math.max(this.lastSelectionStart, this.lastSelectionEnd));
@@ -585,8 +586,8 @@ export class ReplReadline implements EditableDocument {
             }
         }
 
-        this.lastSelectionStart = this.selectionStart;
-        this.lastSelectionEnd = this.selectionEnd;
+        this.lastSelectionStart = this.selectionLeft;
+        this.lastSelectionEnd = this.selectionRight;
 
         this.updateParenMatches()
         this._repaintListeners.forEach(x => x());
@@ -636,8 +637,8 @@ export class ReplReadline implements EditableDocument {
     }
 
     private mouseDrag = (e: MouseEvent) => {
-        this.selectionEnd = this.pageToOffset(e.pageX, e.pageY)
-        this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+        this.selectionRight = this.pageToOffset(e.pageX, e.pageY)
+        this.caretX = this.model.getRowCol(this.selectionRight)[1];
         this.repaint();
     }
 
@@ -649,8 +650,8 @@ export class ReplReadline implements EditableDocument {
     private mouseDown = (e: MouseEvent) => {
         e.preventDefault();
 
-        this.selectionStart = this.selectionEnd = this.pageToOffset(e.pageX, e.pageY)
-        this.caretX = this.model.getRowCol(this.selectionEnd)[1];
+        this.selectionLeft = this.selectionRight = this.pageToOffset(e.pageX, e.pageY)
+        this.caretX = this.model.getRowCol(this.selectionRight)[1];
         this.repaint();
 
         window.addEventListener("mousemove", this.mouseDrag)
@@ -705,7 +706,7 @@ export class ReplReadline implements EditableDocument {
     }
 
     public canReturn() {
-        return this.selectionEnd == this.selectionStart && this.selectionEnd == this.model.maxOffset;
+        return this.selectionRight == this.selectionLeft && this.selectionRight == this.model.maxOffset;
     }
 
     public freeze() {
@@ -716,7 +717,7 @@ export class ReplReadline implements EditableDocument {
         this.wrap.removeEventListener("touchstart", this.focus);
         this.input.disabled = true;
 
-        this.selectionStart = this.selectionEnd = this.model.maxOffset;
+        this.selectionLeft = this.selectionRight = this.model.maxOffset;
         this.repaint();
         this.caret.parentElement.removeChild(this.caret);
     }
@@ -725,7 +726,7 @@ export class ReplReadline implements EditableDocument {
         this.freeze();
     }
 
-    growSelectionStack: { anchor: number, active: number}[] = [];
+    selectionStack: ModelEditSelection[] = [];
 }
 
 /**